+++ proyecto-regenera/proyecto-regenera-main/proyecto-regenera-be/src/main/java/proyecto/repository/AspectoAmbientalTemaRepository.java
@@ -6,11 +6,11 @@ import proyecto.model.AspectoAmbientalTemaModel;
 import java.util.List;
 import java.util.Optional;
 
-public interface AspectoAmbientalTemaRepository extends JpaRepository<AspectoAmbientalTemaModel, Integer> {
+public interface AspectoAmbientalTemaRepository extends JpaRepository<AspectoAmbientalTemaModel, Long> {
 
     @Override
     List<AspectoAmbientalTemaModel> findAll();
 
-    Optional<AspectoAmbientalTemaModel> findById(Integer id);
+    Optional<AspectoAmbientalTemaModel> findById(Long id);
 
 }
