+++ proyecto-regenera/proyecto-regenera-main/proyecto-regenera-be/src/main/java/proyecto/request_response/RequisitoLegalResponse.java
@@ -33,7 +33,7 @@ public class RequisitoLegalResponse {
 
     String resultado;
 
-    Integer idAspecto;
+    Long idAspecto;
 
     LocalDate fecha;
